+++ release/v/vietnam/HISTORY.md
@@ -1,6 +1,16 @@
 vietnam Change History
 ====================
 
+2.0.3 (2025-10-06)
+----------------
+* Updated by cathaylab
+* added tone marks generation rules for "uyu" vowel
+
+2.0.2 (2025-09-12)
+----------------
+* Updated by cathaylab
+* fixing bug of keyboard codes
+
 2.0.1 (2025-07-12)
 ----------------
 * Updated by cathaylab